requirements.txt
Flask
python-dotenv
google-generativeai
Pillow
opencv-python
moviepy
requests
whisper
torch
torchaudio
werkzeug
python-docx
mysql-connector-python